+++ src/main/java/com/hambug/Hambug/domain/user/dto/UserDto.java
@@ -40,6 +40,9 @@ public class UserDto {
     boolean isActive;
 
     @Getter(onMethod_ = {@JsonIgnore, @Schema(hidden = true)})
+    String socialId;
+
+    @Getter(onMethod_ = {@JsonIgnore, @Schema(hidden = true)})
     String accessToken;
 
     @Getter(onMethod_ = {@JsonIgnore, @Schema(hidden = true)})
@@ -71,6 +74,7 @@ public class UserDto {
                 .profileImageUrl(user.getProfileImageUrl())
                 .loginType(user.getLoginType())
                 .email(user.getEmail())
+                .socialId(user.getSocialId())
                 .role(user.getRole())
                 .isActive(user.isActive()).build();
     }
