+++ src/main/java/com/hambug/Hambug/global/constatns/Security.java
@@ -13,7 +13,9 @@ public class Security {
             "/js/**",
             "/images/**",
             "/webjars/**",
-            "/api/v1/auth/**",
+            "/api/v1/auth/refresh",
+            "/api/v1/auth/login/**",
+            "/api/v1/auth/callback",
             "/api/swagger-ui/**",
             "/api/swagger-ui.html",
             "/swagger-ui.html",
